+++ compiler/main.ts
@@ -1,4 +1,4 @@
-import Parser from "./reader/parser";
+import parse from "./reader/parser";
 import { evaluate } from "./runtime/interpreter";
 import Environment from "./runtime/environment";
 import { readFileSync, writeFileSync } from "fs";
@@ -7,7 +7,7 @@ async function main() {
     const environment = new Environment()
 
     const input = String(readFileSync("./sample.txt"))
-    const program = Parser(input)
+    const program = parse(input)
     writeFileSync("./data/ast.json", JSON.stringify(program, null, 4))
     
     console.log("== sample.txt ==")